Do not use int to hold memory count
int overflows after 2GB.
Patch from Wolfgang Lorenz, thanks!

CCBUG:184323

Backported to 4.6 branch.

svn path=/branches/KDE/4.6/kdegraphics/gwenview/; revision=1218970

+++ app/savebar.cpp
@@ -153,8 +153,8 @@ struct SaveBarPrivate {
 
 	void updateTooManyChangesFrame(const QList<KUrl>& list) {
 		qreal maxPercentageOfMemoryUsage = GwenviewConfig::percentageOfMemoryUsageWarning();
-		int maxMemoryUsage = MemoryUtils::getTotalMemory() * maxPercentageOfMemoryUsage;
-		int memoryUsage = 0;
+		qulonglong maxMemoryUsage = MemoryUtils::getTotalMemory() * maxPercentageOfMemoryUsage;
+		qulonglong memoryUsage = 0;
 		Q_FOREACH(const KUrl& url, list) {
 			Document::Ptr doc = DocumentFactory::instance()->load(url);
 			memoryUsage += doc->memoryUsage();
